backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# MongoDB
motor>=3.6.0
pymongo>=4.10.0
beanie>=1.27.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0

# Validation and Settings
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Redis for caching
redis>=5.2.0

# Utilities
python-dateutil>=2.9.0
httpx>=0.28.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Logging
loguru>=0.7.2
